Install temporary /etc/resolv.conf again.
In the new `bootstrap_packages()` from `common.sh`, temporarily install
a copy of the host's /etc/resolv.conf to ensure that we really can
bootstrap `pkg(8)` on slave build nodes. Apparently I was too hasty to
remove this before: it may not be required on the head build node, but
it seems to be required on slave nodes.

This partially reverts the semantics of commit
c06266c8ac3b6988fc2e4889006740a9303cea33, though not syntactically.