Move package bootstrapping to a firstboot script.
Up until now we've been relying on the host environment to run `pkg(8)`
bootstrapping from a `chroot(8)` session. As our target VM images
diverge more from vanilla FreeBSD (especially with respect to metaio),
we can no longer rely on basic utilities like `cp(1)` to work correctly
on the host kernel (which lacks metaio versions of system calls). This
causes problems when package post-install scripts expect base system
utilities to be available (as is true for packages like `kyua` that we
install on almost every VM image).

Our new approach is to install metadata about package requirements into
the VM images (in `/etc/rc.conf`) and then pass this information to a
`firstboot` script in `/etc/rc.d/pkg_bootstrap`. This does require the
VM to have network access in order to bootstrap `pkg(8)` and to install
its packages, but it works more reliably in what is now effectively a
cross-build scenario.