Create VM /etc/passwd using host `pwd_mkdb(8)`.
We aren't able to create a VM `/etc/passwd` using the VM's `pwd_mkdb(8)`
from within a `chroot(8)` environment, but we *can* use the host's
version of the command and pass it the `-d` flag.

This also means that we don't have to create an `/etc/passwd` from
within `/etc/rc.d/pkg_bootstrap`, which didn't help us anyway with tools
like `dhclient(8)` that wanted to drop privileges.