Attempt to unify Jenkins and non-Jenkins builds.
Rewrite the top of this script to separate those aspects of
configuration that depend on the details of Jenkins jobs vs a local
user's setup from those that should be common. Hopefully this will
allow us to test our build scripts in the course of normal development,
not just on the CI systems.